Move 'dbutil' scripts to `seedctl admin database *`
