+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["libseed", "seedctl", "web", "dbutil"]
+members = ["libseed", "seedctl", "web"]
 resolver = "3"
 
 [workspace.package]
