+++ Makefile
@@ -61,7 +61,7 @@ INIT_DB ?= $(SEEDWEB_DATABASE_DIR)/seedcollection.sqlite
 INIT_DB_ARGS ?= --download
 prepare-db: ./db/itis/minnesota-native-status.csv ./db/germination/germination-data.csv
 	cargo run -p seedctl -- admin -d $(INIT_DB) database init $(INIT_DB_ARGS)
-	cargo run -p dbutil -- native-status -d $(INIT_DB) --updatedb ./db/itis/minnesota-native-status.csv
-	cargo run -p dbutil -- germination-codes -d $(INIT_DB) --updatedb ./db/germination/germination-data.csv
+	cargo run -p seedctl -- admin -d $(INIT_DB) database update-native-status ./db/itis/minnesota-native-status.csv
+	cargo run -p seedctl -- admin -d $(INIT_DB) database update-germination-info ./db/germination/germination-data.csv
 
 .PHONY: check-sqlx-env prepare-db container run-container
