--- dbutil/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "dbutil"
-version = { workspace = true }
-edition = { workspace = true }
-
-[dependencies]
-anyhow = { workspace = true }
-sqlx = { workspace = true }
-tokio = { workspace = true }
-csv = "1.3"
-clap = { workspace = true }
-tracing = { workspace = true }
-tracing-subscriber = { workspace = true, features = ["env-filter"] }
-indicatif = { version = "0.17.11" }
